src: move cci_core_types.h to cci_core
