+++ src/cci_cfg/cci_param_untyped.h
@@ -28,7 +28,7 @@
 
 #include "cci_cfg/cci_config_macros.h"
 #include "cci_cfg/cci_param_if.h"
-#include "cci_cfg/cci_core_types.h"
+#include "cci_core/cci_core_types.h"
 #include "cci_cfg/cci_value.h"
 #include "cci_cfg/cci_originator.h"
 #include <greencontrol/config.h>
