+++ src/cci_core/cci_core_types.h
